DataProcessingSite/requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.5
django-crispy-forms==1.14.0
django-tables2==2.5.1
et-xmlfile==1.1.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.1
openpyxl==3.1.0
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
xlwt==1.3.0
